+++ src/modules/baselayerSwitcher/components/BaselayerSwitcher.vue
@@ -19,6 +19,7 @@ export default {
             "activatedExpandable",
             "baselayerIds",
             "configPaths",
+            "singleBaseLayer",
             "topBaselayerId",
             "type"
         ])
@@ -127,6 +128,13 @@ export default {
             selectableBackroundLayerIds.splice(index, 1);
             if (this.topBaselayerId !== null) {
                 selectableBackroundLayerIds.push(this.topBaselayerId);
+                if (this.singleBaseLayer) {
+                    this.layerConfigsByAttributes({
+                        id: this.topBaselayerId
+                    }).forEach(layer => {
+                        layer.visibility = false;
+                    });
+                }
             }
             this.setBaselayerIds(selectableBackroundLayerIds);